added zero crossing, update solver stateful
now the solver does 10 steps to refine the error in zero crossing time by a factor of 1000.
also now we save the state in each and every stateful block, instead of being part of the solver.

+++ src/BlockClasses/ScopeDisplayClass.cpp
@@ -77,8 +77,8 @@ node::BlockClass::GetFunctorResult node::ScopeDisplayClass::GetFunctor(const std
 	assert(properties.size() == 1);
 	assert(std::holds_alternative<uint64_t>(properties[0].prop));
 	uint64_t value = std::get<uint64_t>(properties[0].prop);
-	std::pmr::vector<int64_t> ports;
-	for (uint64_t i = 0; i < value; i++)
+	std::pmr::vector<int32_t> ports;
+	for (uint32_t i = 0; i < value; i++)
 	{
 		ports.push_back(i);
 	}
